Reimplement logo.
Done in OmniGraffle and Sketch. The Sketch file is included in this commit;
the OmniGraffle bit was where I redrew the logo, but I think Sketch is the
better place to make updates in the future. It has configs to export masthead,
bug, and text files in PNG and SVG.

Use the new masthead.svg and .png in the header for the logo.