Expand macOS docs for core homebrew formula
I'm working on a formula for the core Homebrew repository that would
support all but Firebird and Oracle. So highlight it here, and document
the installation of additional libraries. The Homebrew formula will link
to this page when installation finishes.

But also cover the tap for those who need to manage Oracle or Firebird
databases.